somewhat safer typecasting in case sizeof(long) != sizeof(void *) (is there even such platforms?)

+++ lib/hostip.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: hostip.c,v 1.140 2004-04-06 15:09:43 bagder Exp $
+ * $Id: hostip.c,v 1.141 2004-04-12 06:55:25 bagder Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -1012,11 +1012,11 @@ static struct hostent* pack_hostent(char** buf, struct hostent* orig)
 
   /* now, shrink the allocated buffer to the size we actually need, which
      most often is only a fraction of the original alloc */
-  newbuf=(char *)realloc(*buf, (long)bufptr-(long)(*buf));
+  newbuf=(char *)realloc(*buf, (long)(bufptr-*buf));
 
   /* if the alloc moved, we need to adjust things again */
   if(newbuf != *buf)
-    hostcache_fixoffset((struct hostent*)newbuf, (long)newbuf-(long)*buf);
+    hostcache_fixoffset((struct hostent*)newbuf, (long)(newbuf-*buf));
 
   /* setup the return */
   *buf = newbuf;
